+++ packages/data/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./types";
 export * from "./opportunity-index";
+export * from "./opportunity-index-examples";
 export * from "./loaders";
 export * from "./validation";